build: Fix totem-plugin option
The third argument to AC_ARG_ENABLE is action-if-present, it will be used when either
--{enable,disable}-totem-plugin is passed. Using enable_totem="yes" would mean that
the totem plugin is always enabled, so instead we have to use $enableval here to get
what we want.

Also check if we got a valid plugin directory with pkg-config and bail out if we didn't.
Otherwise TOTEM_PLUGIN_DIR would have been empty in the case that --enable-totem-plugin
was passed but totem wasn't installed.